fix: added suspense boundary
