+++ src/components/index.ts
@@ -1,8 +1,23 @@
 import CustomPagination from './CustomPagination'
 import Error from './Error'
 import Header from './Header'
+import Movie from './Movie'
 import MovieCard from './MovieCard'
+import MovieCardSkeletton from './MovieCardSkeletton'
 import MovieInformation from './MovieInformation'
+import MovieInformationSkeletton from './MovieInformationSkeletton'
+import Movies from './Movies'
 import SearchField from './SearchField'
 
-export { CustomPagination, Error, Header, MovieCard, MovieInformation, SearchField }
+export {
+  CustomPagination,
+  Error,
+  Header,
+  Movie,
+  Movies,
+  MovieCard,
+  MovieCardSkeletton,
+  MovieInformation,
+  MovieInformationSkeletton,
+  SearchField,
+}
